ref: Restructure project for single import
