+++ lib/src/_connect_api.dart
@@ -1,7 +1,8 @@
 import 'dart:async';
 
-import 'package:stomp_dart_client/stomp_config.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
+
+import 'stomp_config.dart';
 
 Future<WebSocketChannel> connect(StompConfig config) {
   throw UnsupportedError('No implementation of the connect api provided');
