+++ lib/src/sock_js/sock_js_utils.dart
