+++ lib/src/stomp_exception.dart
