+++ lib/src/stomp_frame.dart
