+++ test/sock_js_parser_test.dart
@@ -1,7 +1,6 @@
 import 'dart:convert';
 
-import 'package:stomp_dart_client/sock_js/sock_js_parser.dart';
-import 'package:stomp_dart_client/stomp_frame.dart';
+import 'package:stomp_dart_client/stomp_dart_client.dart';
 import 'package:test/test.dart';
 
 void main() {
