+++ test/stomp_handler_test.dart
@@ -1,9 +1,6 @@
 import 'dart:async';
 
-import 'package:stomp_dart_client/stomp_config.dart';
-import 'package:stomp_dart_client/stomp_exception.dart';
-import 'package:stomp_dart_client/stomp_frame.dart';
-import 'package:stomp_dart_client/stomp_handler.dart';
+import 'package:stomp_dart_client/stomp_dart_client.dart';
 import 'package:stream_channel/stream_channel.dart';
 import 'package:test/test.dart';
